move PaymentIntentResult to domain
