+++ uisdk/src/main/java/tech/dojo/pay/uisdk/presentation/ui/virtualterminalcheckout/mapper/VirtualTerminalViewEntityMapper.kt
@@ -1,6 +1,6 @@
 package tech.dojo.pay.uisdk.presentation.ui.virtualterminalcheckout.mapper
 
-import tech.dojo.pay.uisdk.data.entities.PaymentIntentResult
+import tech.dojo.pay.uisdk.domain.entities.PaymentIntentResult
 import tech.dojo.pay.uisdk.domain.entities.SupportedCountriesDomainEntity
 import tech.dojo.pay.uisdk.presentation.ui.carddetailscheckout.entity.SupportedCountriesViewEntity
 import tech.dojo.pay.uisdk.presentation.ui.carddetailscheckout.mapper.AllowedPaymentMethodsViewEntityMapper
@@ -19,25 +19,27 @@ import java.util.Currency
 
 internal class VirtualTerminalViewEntityMapper(
     private val supportedCountriesViewEntityMapper: SupportedCountriesViewEntityMapper,
-    private val allowedPaymentMethodsViewEntityMapper: AllowedPaymentMethodsViewEntityMapper
+    private val allowedPaymentMethodsViewEntityMapper: AllowedPaymentMethodsViewEntityMapper,
 ) {
 
     fun apply(
         paymentIntentResult: PaymentIntentResult.Success,
-        supportedCountriesDomainEntity: List<SupportedCountriesDomainEntity>
+        supportedCountriesDomainEntity: List<SupportedCountriesDomainEntity>,
     ) = VirtualTerminalViewState(
         isLoading = false,
         paymentDetailsSection = getPaymentDetailsSectionWithPaymentIntent(
-            paymentIntentResult
+            paymentIntentResult,
         ),
         shippingAddressSection = getShippingAddressSectionWithPaymentIntent(
-            paymentIntentResult, supportedCountriesDomainEntity
+            paymentIntentResult,
+            supportedCountriesDomainEntity,
         ),
         billingAddressSection = getBillingAddressSectionWithPaymentIntent(
-            paymentIntentResult, supportedCountriesDomainEntity
+            paymentIntentResult,
+            supportedCountriesDomainEntity,
         ),
         cardDetailsSection = getCardDetailsSectionWithPaymentIntent(paymentIntentResult),
-        payButtonSection = PayButtonViewState()
+        payButtonSection = PayButtonViewState(),
     )
 
     private fun getPaymentDetailsSectionWithPaymentIntent(paymentIntentResult: PaymentIntentResult.Success) =
@@ -45,12 +47,12 @@ internal class VirtualTerminalViewEntityMapper(
             totalAmount = paymentIntentResult.result.amount.valueString,
             amountCurrency = Currency.getInstance(paymentIntentResult.result.amount.currencyCode).symbol,
             merchantName = paymentIntentResult.result.merchantName,
-            orderId = paymentIntentResult.result.orderId
+            orderId = paymentIntentResult.result.orderId,
         )
 
     private fun getShippingAddressSectionWithPaymentIntent(
         paymentIntentResult: PaymentIntentResult.Success,
-        supportedCountriesDomainEntity: List<SupportedCountriesDomainEntity>
+        supportedCountriesDomainEntity: List<SupportedCountriesDomainEntity>,
     ): ShippingAddressViewState {
         val countryList = getSupportedCountriesList(supportedCountriesDomainEntity)
         val currentSelectedCountry = if (countryList.isNotEmpty()) {
@@ -66,18 +68,20 @@ internal class VirtualTerminalViewEntityMapper(
                 0
             },
             supportedCountriesList = countryList,
-            currentSelectedCountry = currentSelectedCountry
+            currentSelectedCountry = currentSelectedCountry,
         )
     }
 
     private fun getBillingAddressSectionWithPaymentIntent(
         paymentIntentResult: PaymentIntentResult.Success,
-        supportedCountriesDomainEntity: List<SupportedCountriesDomainEntity>
+        supportedCountriesDomainEntity: List<SupportedCountriesDomainEntity>,
     ): BillingAddressViewState {
         val isSectionVisible =
             if (paymentIntentResult.result.collectionShippingAddressRequired && paymentIntentResult.result.collectionBillingAddressRequired) {
                 false
-            } else paymentIntentResult.result.collectionBillingAddressRequired
+            } else {
+                paymentIntentResult.result.collectionBillingAddressRequired
+            }
         val countryList = getSupportedCountriesList(supportedCountriesDomainEntity)
         val currentSelectedCountry = if (countryList.isNotEmpty()) {
             countryList[0]
@@ -107,15 +111,16 @@ internal class VirtualTerminalViewEntityMapper(
                 SECOND_SECTION_WITH_BILLING_OFF_SET_DP
             },
             emailInputField = InputFieldState(
-                value = "", isVisible = paymentIntentResult.result.collectionEmailRequired
+                value = "",
+                isVisible = paymentIntentResult.result.collectionEmailRequired,
             ),
             cardHolderInputField = InputFieldState(value = ""),
             cardNumberInputField = InputFieldState(value = ""),
             cardExpireDateInputField = InputFieldState(value = ""),
             cvvInputFieldState = InputFieldState(value = ""),
             allowedPaymentMethodsIcons = allowedPaymentMethodsViewEntityMapper.apply(
-                paymentIntentResult.result.supportedCardsSchemes
-            )
+                paymentIntentResult.result.supportedCardsSchemes,
+            ),
         )
 
     private fun getSupportedCountriesList(supportedCountriesDomainEntity: List<SupportedCountriesDomainEntity>): List<SupportedCountriesViewEntity> {
