+++ uisdk/src/test/java/tech/dojo/pay/uisdk/presentation/ui/paymentmethodcheckout/viewmodel/PaymentMethodCheckoutViewModelTest.kt
@@ -21,7 +21,6 @@ import tech.dojo.pay.sdk.card.presentation.card.handler.DojoSavedCardPaymentHand
 import tech.dojo.pay.sdk.card.presentation.gpay.handler.DojoGPayHandler
 import tech.dojo.pay.uisdk.R
 import tech.dojo.pay.uisdk.core.MainCoroutineScopeRule
-import tech.dojo.pay.uisdk.data.entities.PaymentIntentResult
 import tech.dojo.pay.uisdk.domain.ObservePaymentIntent
 import tech.dojo.pay.uisdk.domain.ObservePaymentMethods
 import tech.dojo.pay.uisdk.domain.ObservePaymentStatus
@@ -30,6 +29,7 @@ import tech.dojo.pay.uisdk.domain.UpdateWalletState
 import tech.dojo.pay.uisdk.domain.entities.AmountDomainEntity
 import tech.dojo.pay.uisdk.domain.entities.FetchPaymentMethodsResult
 import tech.dojo.pay.uisdk.domain.entities.PaymentIntentDomainEntity
+import tech.dojo.pay.uisdk.domain.entities.PaymentIntentResult
 import tech.dojo.pay.uisdk.domain.entities.PaymentMethodsDomainEntity
 import tech.dojo.pay.uisdk.domain.entities.PaymentMethodsDomainEntityItem
 import tech.dojo.pay.uisdk.presentation.ui.carddetailscheckout.state.InputFieldState
@@ -94,7 +94,9 @@ class PaymentMethodCheckoutViewModelTest {
     @Test
     fun `should emit state with allowedCardNetworks when payment Intent emits  with supportedWalletSchemes contains gpay and gPayConfig not null`() {
         // arrange
-        val paymentIntentFakeFlow: MutableStateFlow<PaymentIntentResult> = MutableStateFlow(PaymentIntentResult.None)
+        val paymentIntentFakeFlow: MutableStateFlow<PaymentIntentResult> = MutableStateFlow(
+            PaymentIntentResult.None,
+        )
         whenever(observePaymentIntent.observePaymentIntent()).thenReturn(paymentIntentFakeFlow)
         paymentIntentFakeFlow.tryEmit(
             PaymentIntentResult.Success(
@@ -155,7 +157,9 @@ class PaymentMethodCheckoutViewModelTest {
     @Test
     fun `test state when gpay is not available from the payment intent `() = runTest {
         // arrange
-        val paymentIntentFakeFlow: MutableStateFlow<PaymentIntentResult> = MutableStateFlow(PaymentIntentResult.None)
+        val paymentIntentFakeFlow: MutableStateFlow<PaymentIntentResult> = MutableStateFlow(
+            PaymentIntentResult.None,
+        )
         whenever(observePaymentIntent.observePaymentIntent()).thenReturn(paymentIntentFakeFlow)
         paymentIntentFakeFlow.tryEmit(
             PaymentIntentResult.Success(
